We will call a word any sequence of one or more letters. A text is a set of words separated by spaces.
Implement a TextHandler class that describes an initially empty extensible set of words. When instantiated, the class must not take any arguments.

An instance of the TextHandler class must have three methods:

add_words() - a method that takes a text as an argument and adds words from the given text to the set
get_shortest_words() - a method that returns an up-to-date list of the shortest words in the set
get_longest_words() - a method that returns an up-to-date list of the longest words in the set
Note 1: The words in the lists returned by the get_shortest_words() and get_longest_words() methods must be in the order in which they were added to the set.

Note 2. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.