Implement a Todo class that describes a to-do list. When instantiated, the class must not take any arguments.
An instance of the Todo class must have one attribute:

things - initially empty to-do list
The Todo class must have four instance methods:

add() - a method that takes the name of the case and its priority (integer) and adds this case to the list of cases as a tuple:
(<case name>, <priority>)
get_by_priority() - a method that takes an integer n as an argument and returns a list of case names with priority n
get_low_priority() - a method that returns a list of case names with the lowest priority
get_high_priority() - a method that returns a list of case names with the highest priority
Note 1: Case names in the lists returned by the get_by_priority(), get_low_priority(), and get_high_priority() methods must appear in the order in which they were added to the list.

Note 2. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.