Implement a Postman class that describes a postman. When instantiated, the class must not take any arguments.
An instance of the Postman class must have one attribute:

delivery_data - initially empty list of addresses to which letters should be delivered
An instance of the Postman class must have three instance methods:

add_delivery() is a method that takes a street, house, and apartment as arguments, and adds this data to the list of addresses as a tuple:
(<street>, <house>, <apartment>)
get_houses_for_street() - a method that takes a street as an argument and returns a list of all the houses on that street that you want to deliver letters to
get_flats_for_house() - a method that takes a street and a house as arguments and returns a list of all apartments in this house to which you want to deliver letters
Note 1: Houses and flats in the lists returned by the get_houses_for_street() and get_flats_for_house() methods must be in the order in which they were added.

Note 2. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.