We will call a word any sequence of one or more Latin letters.
Implement a Wordplay class that describes an extensible set of words. When instantiated, the class must take one argument:

words is a list that defines the initial set of words. If not passed, the initial set of words is considered empty
An instance of the Wordplay class must have one attribute:

words - a list containing a set of words
The Wordplay class must have four instance methods:

add_word() is a method that takes a word as an argument and adds it to the set. If the word is already in the set, the method should do nothing.
words_with_length() - a method that takes a natural number n as an argument and returns a list of words from the set, the length of which is n
only() - a method that takes an arbitrary number of arguments - letters, and returns all words from the set, which include only the passed letters
avoid() - a method that takes an arbitrary number of arguments - letters, and returns all words from the words list that do not contain any of these letters
Note 1: The words in the lists returned by the words_with_length(), only(), and avoid() methods must appear in the order in which they were added.

Note 2: An instance of the Wordplay class must not depend on the list from which it was created. In other words, if the original list changes, then the instance of the Wordplay class should not change.

Note 3. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.