Implement a BankAccount class that describes a bank account. When instantiated, the class must take one argument:
balance — account balance, default value is 0
An instance of the BankAccount class must have one attribute:

_balance — account balance
The BankAccount class must have four instance methods:

get_balance() - method that returns the current account balance
deposit() - a method that takes an amount as an argument and increases the account balance by amount
withdraw() is a method that takes an amount as an argument and reduces the account balance by amount. If amount exceeds the amount of funds in the account balance, a ValueError exception should be raised with the message:
There are not enough funds on the account
transfer() is a method that takes a bank account account and an amount as arguments. The method should decrease the current account balance by amount and increase the account balance by amount. If amount exceeds the amount of funds in the current account balance, a ValueError exception should be raised with the message:
There are not enough funds on the account
Note 1. We will consider instances of classes int and float as numbers.

Note 2. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.