Implement a User class that describes an Internet user. When instantiated, the class must take two arguments in the following order:
name is the username. If name is not a non-empty, all-letter string, a ValueError exception must be raised with the text:
Invalid name
age is the age of the user. If age is not an integer belonging to the interval [0; 110], a ValueError exception should be raised with the text:
Incorrect age
An instance of the User class must have two attributes:

_name - username
_age - user's age
The User class must have four instance methods:

get_name() - method returning the username
set_name() is a method that takes new_name as an argument and changes the username to new_name. If new_name is not a non-empty, all-letter string, a ValueError exception must be raised with the text:
Invalid name
get_age() - method that returns the user's age
set_age() is a method that takes new_age as an argument and changes the user's age to new_age. If new_age is not an integer belonging to the segment [0; 110], a ValueError exception should be raised with the text:
Incorrect age
Note 1: If both the name and age are invalid when an instance of the User class is created, an exception associated with the name should be thrown.