For security purposes, passwords from user accounts are stored in databases not explicitly, but as hash values - numbers calculated using a special algorithm based on passwords.
You have a hash_function() function that takes a password as an argument and returns its hash value.

Implement an Account class that describes an Internet user's account on some service. When instantiated, the class must take two arguments in the following order:

login - user login
password — user password
The Account class must have two properties:

login is a read-only property that returns the user's login. When attempting to change a property, an AttributeError exception must be raised with the text:
Login cannot be changed
password is a read/write property that returns the hash value of the password from the user's account. When changed, the property should calculate the hash value of the new password and save it, not the password itself
Note 1. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.

Note 2. There are no restrictions regarding the implementation of the Account class, it can be arbitrary.