Implement a Color class that describes a color. When instantiated, the class must take one argument:
hexcode - hexadecimal color value
An instance of the Color class must have three attributes:

r is the intensity of the red component of the color as a decimal number
g is the intensity of the green component of the color as a decimal number
b is the intensity of the blue component of the color as a decimal number
The Color class must have one property:

hexcode is a read/write property that returns a hexadecimal color value
Note 1: When changing the hexadecimal color value, the values of the r, g, and b attributes must also change.

Note 2. It is guaranteed that only uppercase Latin letters are used to write hexadecimal numbers.

Note 3. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.

Note 4. There are no restrictions regarding the implementation of the Color class, it can be arbitrary.