Implement a QuadraticPolynomial class that describes a square trinomial. When instantiated, the class must take three arguments in the following order:
a - coefficient

a of the square trinomial
b - coefficient

b of the square trinomial
c - coefficient

c square trinomial
An instance of the QuadraticPolynomial class must have three attributes:

a - coefficient

a of the square trinomial
b - coefficient

b of the square trinomial
c - coefficient

c square trinomial
The QuadraticPolynomial class must have two class methods:

from_iterable() - a method that takes as an argument an iterable object of three elements a, b and c, which represent the coefficients of a square trinomial, and returns an instance of the QuadraticPolynomial class created based on the coefficients passed
from_str() is a method that takes as an argument a string containing the coefficients a, b, and c of the square trinomial, separated by a space. The method must return an instance of the QuadraticPolynomial class, created based on the coefficients passed, previously converted to instances of the float class
Note 1. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.