Implement a Pet class that describes a pet. When instantiated, the class must take one argument:
name - the name of the pet
An instance of the Pet class must have one attribute:

name - the name of the pet
The Pet class must have three class methods:

first_pet() is a method that returns the very first instance of the Pet class created. If no instance has yet been created, the method should return None
last_pet() is a method that returns the most recently created instance of the Pet class. If no instance has yet been created, the method should return None
num_of_pets() - a method that returns the number of created instances of the Pet class
Note 1. There are no restrictions regarding the implementation of the Pet class, it can be arbitrary.

Note 2. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.