Implement the StrExtension class, which describes a set of functions for working with strings. When instantiated, the class must not take any arguments.
The StrExtension class must have three static methods:

remove_vowels() - a method that takes a string as an argument, removes all Latin vowels from it without regard to case and returns the result
leave_alpha() - a method that takes a string as an argument, removes all non-Latin characters from it, and returns the result
replace_all() is a method that takes three string arguments string, chars and char, replaces all characters in string string from chars to char case sensitively and returns the result.
Note 1. It is guaranteed that all alphabetic characters belong to the Latin alphabet.

Note 2. Latin vowels: a, e, i, o, u, y.

Note 3. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.