Snake Case is a style of writing compound words in which several words are separated by an underscore (_) and do not have spaces in the entry, with each word written with a small letter. For example, bee_geek and hello_world.
Upper Camel Case - a style of writing compound words in which several words are written together without spaces, with each word capitalized. For example, BeeGeek and HelloWorld.

Implement the CaseHelper class, which describes a set of functions for working with strings in the Snake Case and Upper Camel Case styles. When instantiated, the class must not take any arguments.

The CaseHelper class must have four static methods:

is_snake() - a method that takes a string as an argument and returns True if the passed string is written in the Snake Case style, or False otherwise
is_upper_camel() - a method that takes a string as an argument and returns True if the passed string is written in the Upper Camel Case style, or False otherwise
to_snake() is a method that takes an Upper Camel Case style string as an argument, writes it in the Snake Case style, and returns the result
to_upper_camel() is a method that takes a string in the Snake Case style as an argument, writes it in the style of the Upper Camel Case and returns the result
Note 1. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.