You have access to the Processor class. When instantiated, the class does not take any arguments.
The Processor class has one static method:

process() is a method that takes an arbitrary object as an argument, converts it depending on its type, and returns the result. If the type of the passed object is not supported by the method, a TypeError exception is raised with the text:
Passed type argument is not supported
Rewrite the Processor class's process() method with the @singledispatchmethod decorator to do the same thing.

Note 1: Examples of object conversions of all supported types are shown in the process() method of the Processor class.

Note 2. There are no restrictions regarding the implementation of the Processor class, it can be arbitrary.