Implement the Negator class. When instantiated, the class must not take any arguments.
The Negator class must have one static method:

neg() is a method that takes an object as an argument and returns its opposite value. If an integer or real number is passed to the method, it must return that number with the opposite sign. If a boolean value is passed as an argument to a method, it must return the opposite boolean value of the one passed. If the passed object is of any other type, a TypeError exception must be raised with the text:
Passed type argument is not supported
Note 1. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.

Note 2. There are no restrictions regarding the implementation of the Negator class, it can be arbitrary.