Implement the Formatter class. When instantiated, the class must not take any arguments.
The Formatter class must have one static method:

format() is a method that takes an object of type int, float, tuple, list or dict as an argument and displays information about the passed object in a format depending on its type. If the passed object is of any other type, a TypeError exception must be raised with the text:
Passed type argument is not supported
Note 1: Examples of formatting objects of all types are shown in the test data.

Note 2: Please note that the format() method must surround the string elements of the collections with apostrophes.

Note 3. Additional data validation is not required. It is guaranteed that the implemented class is only used with valid data.

Note 4. There are no restrictions regarding the implementation of the Formatter class, it can be arbitrary.