Implement a BirthInfo class that describes date of birth data. When instantiated, the class must take one argument:
birth_date is the date of birth, represented as one of the following:
date class instance
date string in ISO format
list or tuple of three integers: year, month and day
If the date of birth is invalid or in some other format, a TypeError exception should be raised with the text:

Passed type argument is not supported
An instance of the BirthInfo class must have one attribute:

birth_date - the date of birth as an instance of the date class
The BirthInfo class must have one property:

age is a read-only property that returns the current age in years, i.e. the number of completed years from the date of birth to today
Note 1: Age in years should be calculated in the same way as the normal age of a person, that is, on the day of birth, his age is increased by one year.